+++ src/main/java/com/api/ecoshieldwebservice/interfaces/IBlogServices.java
@@ -1,15 +1,15 @@
 package com.api.ecoshieldwebservice.interfaces;
 
-import com.api.ecoshieldwebservice.dtos.BlogRequestDTO;
+import com.api.ecoshieldwebservice.dtos.request.BlogRequestDTO;
 import com.api.ecoshieldwebservice.dtos.BlogResponseDTO;
 
 import java.util.List;
 
 public interface IBlogServices {
-    BlogResponseDTO findById(Integer id);
+    BlogResponseDTO findById(Long id);
     BlogResponseDTO registrar(BlogRequestDTO blogRequestDTO);
-    BlogResponseDTO actualizar(Integer id, BlogRequestDTO blogRequestDTO);
-    void borrar(Integer id);
+    BlogResponseDTO actualizar(Long id, BlogRequestDTO blogRequestDTO);
+    void borrar(Long id);
     List<BlogResponseDTO> findAll();
     BlogResponseDTO findTipDelDia();
     List<BlogResponseDTO> findAllNews();
