+++ src/main/java/com/api/ecoshieldwebservice/interfaces/IUsuarioServices.java
@@ -5,7 +5,7 @@ import com.api.ecoshieldwebservice.dtos.*;
 import java.util.List;
 
 public interface IUsuarioServices {
-    public UsuarioProfileDTO findById(Integer id);
-    public UsuarioProfileDTO updateProfile(Integer id, UsuarioProfileDTO usuarioProfileDTO);
+    public UsuarioProfileDTO findById(Long id);
+    public UsuarioProfileDTO updateProfile(Long id, UsuarioProfileDTO usuarioProfileDTO);
     public List<UsuarioResponseDTO> findAll();
 }
